refactor: drop `.../safe` entrypoints (#8823)
