+++ .github/workflows/test-and-release.yml
@@ -126,13 +126,6 @@ jobs:
         node -e '(async () => { await import("@zwave-js/serial") })()'
         node -e '(async () => { await import("@zwave-js/shared") })()'
         node -e '(async () => { await import("@zwave-js/testing") })()'
-        node -e '(async () => { await import("zwave-js/safe") })()'
-        node -e '(async () => { await import("@zwave-js/cc/safe") })()'
-        node -e '(async () => { await import("@zwave-js/config/safe") })()'
-        node -e '(async () => { await import("@zwave-js/core/safe") })()'
-        node -e '(async () => { await import("@zwave-js/nvmedit/safe") })()'
-        node -e '(async () => { await import("@zwave-js/serial/safe") })()'
-        node -e '(async () => { await import("@zwave-js/shared/safe") })()'
         node -e '(async () => { await import("@zwave-js/bindings-browser/fs") })()'
 
         echo "Testing version exports..."
