--- packages/cc/src/index_safe.ts
@@ -1,5 +1,0 @@
-/* @forbiddenImports external */
-
-export * from "./lib/SetValueResult.js";
-// eslint-disable-next-line @zwave-js/no-forbidden-imports -- FIXME: This is actually wrong, but I need to get the release done
-export * from "./lib/_Types.js";
