+++ packages/config/src/devices/DeviceMetadata.ts
@@ -1,6 +1,6 @@
 import { type JSONObject, pick } from "@zwave-js/shared";
 import { isArray, isObject } from "alcalzone-shared/typeguards";
-import { throwInvalidConfig } from "../utils_safe.js";
+import { throwInvalidConfig } from "../utils.js";
 import {
 	type ConditionalItem,
 	conditionApplies,
