+++ packages/config/src/devices/ParamInformation.ts
@@ -7,7 +7,7 @@ import {
 	pick,
 } from "@zwave-js/shared";
 import { isArray, isObject } from "alcalzone-shared/typeguards";
-import { throwInvalidConfig } from "../utils_safe.js";
+import { throwInvalidConfig } from "../utils.js";
 import {
 	type ConditionalItem,
 	conditionApplies,
