+++ packages/config/src/devices/SceneConfig.ts
@@ -1,5 +1,5 @@
 import { type JSONObject, pick } from "@zwave-js/shared";
-import { throwInvalidConfig } from "../utils_safe.js";
+import { throwInvalidConfig } from "../utils.js";
 import {
 	type ConditionalItem,
 	conditionApplies,
