--- packages/config/src/index_safe.ts
@@ -1,5 +1,0 @@
-/* @forbiddenImports external */
-
-export * from "./Logger_safe.js";
-export { PACKAGE_VERSION } from "./_version.js";
-export type * from "./traits.js";
