--- packages/core/src/index_safe.ts
@@ -1,29 +1,0 @@
-/* @forbiddenImports external */
-
-export * from "./definitions/index.js";
-export * from "./dsk/index.js";
-export * from "./error/ZWaveError.js";
-export * from "./log/shared.js";
-// eslint-disable-next-line @zwave-js/no-forbidden-imports -- FIXME: We know this import is safe, but the lint rule doesn't
-export * from "./qr/index.js";
-export * from "./registries/DeviceClasses.js";
-export * from "./registries/Indicators.js";
-export * from "./registries/Meters.js";
-export * from "./registries/Notifications.js";
-export * from "./registries/Scales.js";
-export * from "./registries/Sensors.js";
-export type * from "./traits/index.js";
-export type * from "./util/_Types.js";
-export { deflateSync, gunzipSync, inflateSync } from "./util/compression.js";
-export * from "./util/config.js";
-export * from "./util/crc.js";
-export * from "./util/graph.js";
-export * from "./util/misc.js";
-export * from "./values/Cache.js";
-export * from "./values/CacheBackedMap.js";
-export * from "./values/Duration.js";
-export * from "./values/Metadata.js";
-export * from "./values/Primitive.js";
-export * from "./values/Timeout.js";
-export * from "./values/ValueDB.js";
-export type * from "./values/_Types.js";
