--- packages/serial/src/index_safe.ts
@@ -1,7 +1,0 @@
-/* @forbiddenImports external */
-
-export type { SerialLogContext } from "./log/Logger_safe.js";
-export * from "./message/Constants.js";
-export * from "./message/MessageHeaders.js";
-export * from "./message/SuccessIndicator.js";
-export type * from "./serialport/Bindings.js";
