+++ packages/shared/package.json
@@ -13,10 +13,6 @@
       "browser": "./build/index_browser.js",
       "default": "./build/index.js"
     },
-    "./safe": {
-      "@@dev": "./src/index_safe.ts",
-      "default": "./build/index_safe.js"
-    },
     "./bindings": {
       "@@dev": "./src/bindings.ts",
       "default": "./build/bindings.js"
