--- packages/shared/src/index_safe.ts
@@ -1,23 +1,0 @@
-/* @forbiddenImports external */
-
-export * from "./AsyncQueue.js";
-export { Bytes, type BytesView } from "./Bytes.js";
-export * from "./EventTarget.js";
-export { ObjectKeyMap } from "./ObjectKeyMap.js";
-export type { ReadonlyObjectKeyMap } from "./ObjectKeyMap.js";
-export * from "./ThrowingMap.js";
-export * from "./TimedExpectation.js";
-export * from "./errors.js";
-export * from "./inheritance.js";
-export * from "./strings.js";
-export type * from "./types.js";
-export {
-	areUint8ArraysEqual,
-	assertUint8Array,
-	hexToUint8Array,
-	isUint8Array,
-	uint8ArrayToHex,
-	uint8ArrayToString,
-} from "./uint8array-extras.js";
-export * from "./utils.js";
-export * from "./wrappingCounter.js";
