--- packages/zwave-js/src/Controller_safe.ts
@@ -1,18 +1,0 @@
-export {
-	ProtocolDataRate,
-	RFRegion,
-	RssiError,
-	isRssiError,
-} from "@zwave-js/core";
-export type { RSSI, TXReport } from "@zwave-js/core";
-export type { ZWaveLibraryTypes } from "@zwave-js/core";
-export type { ControllerStatistics } from "./lib/controller/ControllerStatistics.js";
-export { ZWaveFeature } from "./lib/controller/Features.js";
-export * from "./lib/controller/Inclusion.js";
-export type {
-	FirmwareUpdateDeviceID,
-	GetFirmwareUpdatesOptions,
-	RebuildRoutesOptions,
-	RebuildRoutesStatus,
-	SDKVersion,
-} from "./lib/controller/_Types.js";
