--- packages/zwave-js/src/Node_safe.ts
@@ -1,16 +1,0 @@
-export {
-	NODE_ID_BROADCAST,
-	NODE_ID_BROADCAST_LR,
-	NODE_ID_MAX,
-	NodeType,
-	ProtocolDataRate,
-	ProtocolVersion,
-} from "@zwave-js/core";
-export type { DataRate, FLiRS } from "@zwave-js/core";
-export { DeviceClass } from "./lib/node/DeviceClass.js";
-export type {
-	NodeStatistics,
-	RouteStatistics,
-} from "./lib/node/NodeStatistics.js";
-export type { VirtualValueID } from "./lib/node/VirtualNode.js";
-export * from "./lib/node/_Types.js";
