--- packages/zwave-js/src/Utils_safe.ts
@@ -1,24 +1,0 @@
-export {
-	ProtocolDataRate,
-	ProtocolType,
-	ProtocolVersion,
-	Protocols,
-	RouteProtocolDataRate,
-	protocolDataRateToString,
-	rssiToString,
-} from "@zwave-js/core";
-export type { Firmware, FirmwareFileFormat } from "@zwave-js/core";
-export {
-	buffer2hex,
-	formatId,
-	getEnumMemberName,
-	num2hex,
-} from "@zwave-js/shared";
-export { driverPresets } from "./lib/driver/ZWaveOptions.js";
-export {
-	formatLifelineHealthCheckRound,
-	formatLifelineHealthCheckSummary,
-	formatRouteHealthCheckRound,
-	formatRouteHealthCheckSummary,
-	healthCheckRatingToWord,
-} from "./lib/node/HealthCheck.js";
