--- packages/zwave-js/src/Zniffer_safe.ts
@@ -1,7 +1,0 @@
-export { BeamingInfo, MPDUHeaderType } from "@zwave-js/core";
-export type { ZnifferOptions } from "./lib/zniffer/Zniffer.js";
-export {
-	ExplorerFrameCommand,
-	LongRangeFrameType,
-	ZWaveFrameType,
-} from "./lib/zniffer/_Types.js";
