--- packages/zwave-js/src/index_safe.ts
@@ -1,10 +1,0 @@
-/* @forbiddenImports external */
-
-// eslint-disable-next-line @zwave-js/no-forbidden-imports -- FIXME: This is actually wrong, but I need to get the release done
-export * from "./Controller_safe.js";
-// export * from "./Driver";
-export * from "./Error.js";
-export * from "./Node_safe.js";
-export * from "./Utils_safe.js";
-export * from "./Values.js";
-export * from "./Zniffer_safe.js";
